+++ CHANGELOG.md
@@ -66,3 +66,4 @@
     * refactor authentication buttons
     * remove unneeded variable from toMap functions
     * remove need for toMap functions all together
+    * remove register object and transfer proposal to proposalcard